requirements.txt
asgiref==3.2.10
Django==3.1
django-crispy-forms==1.9.2
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-jsonapi==3.1.0
et-xmlfile==1.0.1
inflection==0.5.0
jdcal==1.4.1
openpyxl==3.0.5
psycopg2-binary==2.8.5
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
